+++ lib/src/brambl/builders/locks/lock_template.dart
@@ -1,6 +1,5 @@
 import 'package:topl_common/proto/brambl/models/box/challenge.pb.dart';
 import 'package:topl_common/proto/brambl/models/box/lock.pb.dart';
-import 'package:topl_common/proto/quivr/models/proposition.pb.dart';
 import 'package:topl_common/proto/quivr/models/shared.pb.dart';
 
 import '../../../common/functional/either.dart';
@@ -36,9 +35,8 @@ class PredicateTemplate implements LockTemplate {
 
   factory PredicateTemplate.fromJson(Map<String, dynamic> json) {
     final threshold = json['threshold'] as int;
-    final innerTemplates = (json['innerTemplates'] as List<dynamic>)
-        .map((e) => PropositionTemplate.fromJson(e))
-        .toList();
+    final innerTemplates =
+        (json['innerTemplates'] as List<dynamic>).map((e) => PropositionTemplate.fromJson(e)).toList();
     return PredicateTemplate(innerTemplates, threshold);
   }
   List<PropositionTemplate> innerTemplates;
@@ -46,10 +44,9 @@ class PredicateTemplate implements LockTemplate {
 
   @override
   Either<BuilderError, Lock> build(List<VerificationKey> entityVks) {
-    final result =
-        ThresholdTemplate(innerTemplates, threshold).build(entityVks);
+    final result = ThresholdTemplate(innerTemplates, threshold).build(entityVks);
     return result.flatMap((ip) {
-      if (ip is Proposition_Threshold) {
+      if (ip.hasThreshold()) {
         final innerPropositions = ip.threshold.challenges;
         return Either.right(Lock(
           predicate: Lock_Predicate(
@@ -59,8 +56,7 @@ class PredicateTemplate implements LockTemplate {
               threshold: threshold),
         ));
       } else {
-        return Either.left(BuilderError(
-            'Unexpected inner proposition type: ${ip.runtimeType}'));
+        return Either.left(UnableToBuildPropositionTemplate('Unexpected inner proposition type: ${ip.runtimeType}'));
       }
     });
   }