+++ CHANGELOG.md
@@ -1,8 +1,16 @@
+## 2.0.0-alpha.3
+- [TSDK-630] Bugfix for Language dictionary loading error
+- [TSDK-651] Fixes type check causing template builder to throw errors because of an unexpected inner proposition
+- [TSDK-615] [TSDK-653] Bugfixes for lock's and extra tests
+- [TSDK-660] [TSDK-656] Fixes bug where extended_ed25519 key returned size of 31 instead of 32
+
+
 ## 2.0.0-alpha.2
 - setup linter for better code quality
 - WalletEntity has been renamed to WalletFellowship, Party has been renamed to Fellowship as well
 - Added and optomized exports for missing brambl_dart functionality
 - Made api signature changes for servicekit alpha 1
+
 
 ## 2.0.0-alpha.1
 - Initial version